ActiveSupport::Concernの実装
